+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The types of changes are:
 ### Added
 - Added model and data migrations and CRUD-layer operations for property-specific messaging [#4901](https://github.com/ethyca/fides/pull/4901)
 - Added option in FidesJS SDK to only disable notice-served API [#4965](https://github.com/ethyca/fides/pull/4965)
+- Added access and erasure support for the Greenhouse Harvest integration [#4945](https://github.com/ethyca/fides/pull/4945)
 
 ### Changed
 - Move new data map reporting table out of beta and remove old table from Data Lineage map. [#4963](https://github.com/ethyca/fides/pull/4963)
@@ -36,7 +37,6 @@ The types of changes are:
 ### Added
 - Deprecate LastServedNotice (lastservednoticev2) table [#4910](https://github.com/ethyca/fides/pull/4910)
 - Added erasure support to the Recurly integration [#4891](https://github.com/ethyca/fides/pull/4891)
-- Added access and erasure support to the Greenhouse Harvest integration [#4945](https://github.com/ethyca/fides/pull/4945)
 - Added UI for configuring integrations for detection/discovery [#4922](https://github.com/ethyca/fides/pull/4922)
 - New queue for saving privacy preferences/notices served [#4931](https://github.com/ethyca/fides/pull/4931)
 - Expose number of tasks in queue in worker health check [#4931](https://github.com/ethyca/fides/pull/4931)